% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frdboot.R
\name{frdboot}
\alias{frdboot}
\title{frdboot}
\usage{
frdboot(y, t, x, h, b, cluster = NULL, a = 0.05, Nbc = 500,
  Nci = 999, p = 1, q = 2, kernel = "tri", residual = "HC3")
}
\arguments{
\item{y}{the outcome variable.}

\item{t}{the treatment variable.}

\item{x}{the running variable.}

\item{h}{the bandwidth for estimation.}

\item{b}{the bandwidth for bias correction.}

\item{cluster}{the cluster variable. Default value is NULL.}

\item{a}{the significance level. Default value is 0.05.}

\item{Nbc}{the number of bootstrap replications for bias correction. Default value is 500.}

\item{Nci}{the number of bootstrap replications for confidence interval. Default value is 999.}

\item{p}{the polynomial order for estimation. Default value is 1.}

\item{q}{the polynomial order for bias correction. Default value is 2.}

\item{kernel}{the kernel function: epa, uni or tri. Default value is "tri".}

\item{residual}{the residual adjustment used in wild bootstrap: HC0, HC2 or HC3. Default value is "HC3".}
}
\value{
The original estimator, the bias-corrected estimator and the robust confidence interval.
}
\description{
A function to calculate bias-corrected estimator and its confidence interval in fuzzy regression discontinuity design.
}
